Avoid exception on iOS: separate mutation, enumeration of advertisement dict
Working on a Cordova application that needs to read individual bytes from a
BLE advertisement, I found that on iOS I would occasionally get an uncaught
NSException in the first for loop in serializableAdvertisementData().

I didn't know any Objective-C, but my colleague Asa Freedman could tell
what the for loop was doing, and that it would succeed when serviceData
had no more than one key, otherwise throwing NSException.

Here's a one-liner fix to enumerate a copy of the keys instead of the
object directly.

Signed-off-by: Lebbeous Fogle-Weekley <lebbeous@gmail.com>