More explicitly handle edge cases when starting/stopping notifications on iOS #577
This will now signal an error to the start and stop callbacks if the
characteristic state does not match what is needed for the callback.

Theoretically, start and stop callbacks shouldn't exist concurrently, but if they
somehow end up doing so, each will get notified if the notification state isn't as expected.